+++ app/_lib/interfaces/stitch-settings.ts
@@ -1,0 +1,7 @@
+import { EdgeInsets } from "./edge-insets";
+
+export interface StitchSettings {
+  columnCount: number;
+  edgeInsets: EdgeInsets;
+  pageRange: string;
+}